created task_status PENDING and revision_status APPROVED test on assignments
